📦 refactor: break circular dependency between constants and infra
- constants/base-url.ts and activation-model/model.ts imported webserver from the @/infra barrel, which re-exports files that depend on @/models, which in turn depends on @/constants — a circular import graph, with APP_BASE_URL/API_BASE_URL computed eagerly at module top-level in base-url.ts
- Both now import webserver directly from @/infra/webserver, which only depends on infra/errors and has no path back to constants/models
- Removed webserver from the infra barrel since nothing consumes it through there anymore
- Verified with npx madge --circular: no circular dependencies remain in src
