+++ src/infra/index.ts
@@ -5,4 +5,3 @@ export * from './email'
 export * from './errors'
 export * from './gated'
 export * from './session'
-export * from './webserver'
